% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate.R
\name{allocation_schemes}
\alias{allocation_schemes}
\title{Available allocation schemes}
\usage{
allocation_schemes()
}
\value{
A character vector of scheme names accepted by
\code{\link[=allocate]{allocate()}}.
}
\description{
Available allocation schemes
}
\examples{
allocation_schemes()
}
